Split shell artifact export tests
